[SID:3379] fix(ci): destructive_schema shard 빨강 — 순수함수 테스트를 별 파일로 분리+weights 등재
원인(run 34139079441, shard 1) — test_3379_doc_nudge_default_silence.py가 destructive_
schema 마커 테스트(실 PG 5건)와 마커 없는 순수함수 테스트(10건)를 한 파일에 섞어 뒀다.
CI의 destructive-schema shard 러너는 파일당 `uv run pytest -q "$f"`를 `-m` 필터 없이
돌린다(파일 하나=순수 destructive라는 전제) — 섞인 파일을 만나면 conftest.py의 혼합
수집 차단 가드(story #3186/8236bbc3)가 정상 발동해 그 파일 전체가 "no tests ran"으로
실패한다. 별개로 "destructive schema weights registered" 가드도 이 신규 파일이
shard-weights.json에 미등재라고 잡았다(story 23bf1913).

처방:
- 순수함수(_message_signals_doc_still_being_revised) 표본 10건을 test_3379_doc_nudge_
  signal_words.py(마커 없음, DB 접근 0)로 분리.
- test_3379_doc_nudge_default_silence.py는 이제 destructive_schema 5건만(모듈 전체
  pytestmark, 개별 데코레이터 관례 제거) — 파일 하나=순수 destructive 전제와 정합.
- infra/destructive-schema-shard-weights.json에 신규 파일 등재(가드가 출력한 평균
  폴백 14.5s 그대로, 실측 재확認 필요 라벨).

로컬 재검증: signal_words 10/10(non-destructive) · default_silence 5/5(destructive,
alembic-migrated 신선 DB) · lint_destructive_schema_weights_registered.py가 이 파일을
더 이상 안 잡음(기존에 이미 미등재였던 무관 10개 파일은 범위 밖 — 별건).
