+++ infra/destructive-schema-shard-weights.json
@@ -1,6 +1,5 @@
 {
   "_snapshot_policy": "story #3558(2026-09-06) — CI 실측 전수 재측정(run 34011626732, 8개 shard-durations 산출물 병합, DB 준비 제외 순수 pytest wall time). 이전 스냅샷(story #3383, 로컬×6 추정)을 대체 — 로컬 추정이 아니라 실 CI 값이라 재측정 기준(a)(b)(c)는 그대로 두되 근거가 더 강해졌다: (a) discover 파일 수가 이 스냅샷 대비 +20% 이상, (b) 샤드 간 CI 벽시계가 1.5배 이상 벌어짐, (c) 25분 천장 대비 여유가 다시 좁아짐. story #3558의 shard-durations 산출물+경고 대조가 이 스냅샷의 노후화를 지속 감시한다.",
-  "_drift_remeasure_procedure": "story #3642(2026-09-07, PO 確定) — 3396 러너 정규화 가드 초과가 반복될 때(특히 3396 W×3 여유축·3642 drift 3-run 스트릭 ::warning이 뜬 뒤) 개별 파일을 재측정하는 절차. 전수 재측정(#3558, measured_at)과 달리 이건 «몇 개 파일만» 좁게 고친다. ① gh api runs/{run_id}로 run_attempt 확認 — required 잡이 최초 시도에서 실패/타임아웃하면 GitHub이 자동 재시도해 그 run의 최신 attempt는 성공한 «깨끗한» 로그로 덮인다(gh run view --log가 항상 최신 attempt만 보여준다) — 원인이 된 실사고 숫자는 attempt=1(또는 실패한 attempt)의 job에서만 보인다: `gh api repos/{owner}/{repo}/actions/runs/{run_id}/attempts/{n}/jobs`로 그 attempt의 job id를 찾고 `gh api repos/{owner}/{repo}/actions/jobs/{job_id}/logs`로 원문을 받는다. ② 로그에서 `러너 정규화 판정선: X초(... 중앙값 배율 ×R ...)` 줄로 그 run의 median 배율 R을 읽고, 문제 파일의 `elapsed: Es` 를 `E/R`로 나눠 **러너 정규화값**(그 러너가 정상 속도였다면 걸렸을 시간)을 낸다. ③ 서로 다른 run(가급적 2개 이상, timeout으로 cancel된 run은 median 표본이 작아 편향될 수 있어 참고만) 정규화값을 평균해 새 weight로 등재 — `source`에 각 run/attempt/job id와 계산식을 남긴다(재현 가능하게). ④ 갱신 뒤 `partition()` 결과로 샤드 총합이 여전히 균형인지(#3642 test_shard_balance_stays_even_after_remeasure) 확認 — 그리디 LPT가 weight 합 기준이라 보통 자동으로 풀린다. 주기: 고정 스케줄 없음 — 3642 drift ::warning(3-run 연속)이 뜨거나 3396 가드가 반복 걸리는 파일이 보이면 그때.",
   "measured_at": "2026-09-06",
   "files": [
     {
@@ -25,8 +24,8 @@
     },
     {
       "file": "tests/test_2813_gate_github_check_realdb.py",
-      "sec": 79.0,
-      "source": "story-3642-drift-remeasure(2026-09-07, PO 실사고 3건 재실측 — 각 run의 러너 정규화 median 배율로 나눈 정규화값 평균: run 34103152060 attempt1 job 101682050718(131s÷median×1.71≈76.6s)·run 34112978612 attempt1 job 101713259910(162s÷median×2.00≈81.0s) 두 완주 run 평균 79s. run 34112876593 attempt1 job 101712947116(160s)은 30분 timeout으로 cancel돼 median 표본이 작아(29개, 조기절단 편향) 참고만 하고 평균엔 미반영. 기존 45.0s는 2026-09-06 정상부하 run 1건뿐이라(story-3558) 느린 러너 날의 실제 무게를 못 담았다."
+      "sec": 45.0,
+      "source": "story-3558-run-34011626732(2026-09-06, 부하 정상 run 1건 실측 — 잠정값. actions/upload-artifact shard-durations-*·DB 준비 제외 순수 pytest wall time. 기존 로컬×6 추정 스냅샷을 대체)"
     },
     {
       "file": "tests/test_3360_site_posts.py",
@@ -350,8 +349,8 @@
     },
     {
       "file": "tests/test_3516_channel_post_comments.py",
-      "sec": 55.0,
-      "source": "story-3642-drift-remeasure(2026-09-07, PO 실사고 재실측 — 정규화값 평균: run 34103152060 attempt1 job 101682050718(68s÷median×1.71≈39.8s)·run 34112978612 attempt1 job 101713259910(142s÷median×2.00≈71.0s) 평균 55s. 두 run 편차가 커(39.8~71.0s) 다음 3 run 실측으로 재확認 필요(AC1) — run 34112876593(154s, cancel로 median 미확定)도 참고."
+      "sec": 33.0,
+      "source": "story-3558-run-34011626732(2026-09-06, 부하 정상 run 1건 실측 — 잠정값. actions/upload-artifact shard-durations-*·DB 준비 제외 순수 pytest wall time. 기존 로컬×6 추정 스냅샷을 대체)"
     },
     {
       "file": "tests/test_3516_comment_reply.py",
@@ -1059,11 +1058,6 @@
       "source": "story-3558-run-34011626732(2026-09-06, 부하 정상 run 1건 실측 — 잠정값. actions/upload-artifact shard-durations-*·DB 준비 제외 순수 pytest wall time. 기존 로컬×6 추정 스냅샷을 대체)"
     },
     {
-      "file": "tests/test_3631_role_templates_content_group.py",
-      "sec": 1.0,
-      "source": "story-3631(2026-09-07, 로컬 실측 — test_2648과 동형 규모(role_templates 임시 테이블+마이그 upgrade/downgrade 2회 왕복), 0.21s wall — 인접 sibling(test_2635/2648)과 같은 자릿수로 등재)"
-    },
-    {
       "file": "tests/test_2637_preset_block_template_seed.py",
       "sec": 1.0,
       "source": "story-3558-run-34011626732(2026-09-06, 부하 정상 run 1건 실측 — 잠정값. actions/upload-artifact shard-durations-*·DB 준비 제외 순수 pytest wall time. 기존 로컬×6 추정 스냅샷을 대체)"
@@ -1145,8 +1139,8 @@
     },
     {
       "file": "tests/test_3414_publication_command_cron_retry.py",
-      "sec": 33.0,
-      "source": "story-3642-drift-remeasure(2026-09-07 — 3396 W×2 초과 파일 전수 스윕 중 발견, 2 run 연속 초과: run 34103152060 attempt1(46s÷median×1.71≈26.9s)·run 34112978612 attempt1(79s÷median×2.00≈39.5s) 평균 33s. 기존 13.0s는 story-3562 분할 시 «테스트 개수 비중» 배분값이라 실측이 아니었다(잠정값 딱지 그대로였던 자리)."
+      "sec": 13.0,
+      "source": "story-3562-split-3414-publication-command(2026-09-06, 원본 22건·실측 40.0s(run 34011626732)를 3-way 분할 후 로컬 raw pytest 비중(cron_retry 10건 7.57s/총 24.81s ≈ 31%)으로 배분한 잠정값, 실 CI 샤드 로그로 재확認 필요)"
     },
     {
       "file": "tests/test_3414_publication_command_edges.py",
@@ -1250,13 +1244,8 @@
     },
     {
       "file": "tests/test_3498_generation_budget.py",
-      "sec": 52.0,
-      "source": "story-3619(2026-09-07) — #3971 run 34089668525 shard 4에서 이 파일(당시 25테스트·등재 42s) 격리실행이 100s>88s로 60초 가드 초과. 즉시 rerun(같은 run) 40.30s로 재현이 러너 경합에 크게 좌우됨을 확認. 조각①/⑤/PUT3종(11테스트)을 test_3498_generation_budget_evidence_and_config.py로 분리(원인 실측·분리 판단은 그 파일 docstring) — 이 파일에 남은 12/25 비율로 100s를 안분(12/25*100=52.0)해 등재, 경합 재발까지 감안한 보수값(40.30s 기준 안분 21.0s가 아니라 100s 기준 안분을 택함)"
-    },
-    {
-      "file": "tests/test_3498_generation_budget_evidence_and_config.py",
-      "sec": 48.0,
-      "source": "story-3619(2026-09-07) — test_3498_generation_budget.py 분리 신설분(조각①·⑤·PUT3종, 11테스트). #3971 run 34089668525 100s(25테스트)를 11/25 비율로 안분(11/25*100=44.0, 반올림 여유 48.0) — 로컬 격리측정 7.28s(무경합)와 별개로 CI 경합 배율을 그대로 반영"
+      "sec": 42.0,
+      "source": "story-3558-run-34011626732(2026-09-06, 부하 정상 run 1건 실측 — 잠정값. actions/upload-artifact shard-durations-*·DB 준비 제외 순수 pytest wall time. 기존 로컬×6 추정 스냅샷을 대체)"
     },
     {
       "file": "tests/test_3506_utm_attribution.py",
@@ -1265,8 +1254,8 @@
     },
     {
       "file": "tests/test_3502_insights_board.py",
-      "sec": 22.0,
-      "source": "story-3604-insights-board-durations(PR#3959 CI 실측 — run 34078092461 job 101608057452, elapsed 22s. 로컬 격리측정 6.65~6.68s×2배 잠정값 14s보다 CI가 더 느렸다 — CPU-bound app.main import 없는 파일도 CI 배율이 로컬 대비 3x대. 실측치로 갱신)"
+      "sec": 14.0,
+      "source": "story-3604-insights-board-durations(원인 실측: from app.main import app 1회 비용 ~3.4s(로컬 무경합)가 이 파일에 섞여 있어 등재 39s 대비 경합 시 80s까지 늘던 원인 — 조각②(HTTP 라우터 6개 테스트, app.main 소비)를 test_3502_insights_board_endpoints.py로 분리. 남은 14개는 list_insights_board를 서비스 함수로 직접 부르며 app.main 무관 — 격리 uv run pytest 프로세스(ci.yml per-file 루프와 동형) 2회 로컬 실측 6.65~6.68s, 2배 안전마진 반영한 14s 잠정값, 실 CI 샤드 로그로 재확認 필요)"
     },
     {
       "file": "tests/test_3510_member_role_sync.py",
@@ -1465,8 +1454,8 @@
     },
     {
       "file": "tests/test_3502_insights_board_endpoints.py",
-      "sec": 25.0,
-      "source": "story-3604-insights-board-durations(PR#3959 CI 실측 — run 34078092461 job 101608057407, elapsed 25s. 로컬 격리측정 4.63~4.85s×2배 잠정값 10s보다 CI가 더 느렸다. 실측치로 갱신)"
+      "sec": 10.0,
+      "source": "story-3604-insights-board-durations(신규 분리 파일 — story #3502 조각② HTTP 라우터 6개 테스트, from app.main import app 1회 비용 ~3.4s 포함. 격리 uv run pytest 프로세스 2회 로컬 실측 4.63~4.85s, 2배 안전마진 반영한 10s 잠정값, 실 CI 샤드 로그로 재확認 필요)"
     },
     {
       "file": "tests/test_3603_connection_last_error.py",
@@ -1512,6 +1501,11 @@
       "file": "tests/test_3645_evidence_asset_hook_keys.py",
       "sec": 41.0,
       "source": "story-3645-evidence-asset-hook-keys(PR #4002 CI 가드 발견 — 로컬 raw pytest 11건 6.81s 실측치를 CI 배율(~6x, story #3383) 보수적으로 반영한 41s 잠정값, 실 CI 샤드 로그로 재확認 필요)"
+    },
+    {
+      "file": "tests/test_3379_doc_nudge_default_silence.py",
+      "sec": 14.5,
+      "source": "story-3379-doc-nudge-default-silence(신규 파일 — 5건, PR #4011 CI \"destructive schema weights registered\" 가드가 미등재로 잡으며 출력한 평균 폴백값(14.5s) 그대로 등재, 실 CI 샤드 로그로 재확認 필요)"
     }
   ]
 }
